[clean] create chatWrapper component which stores chat page states and create a context hook which provide this data to chat page and page components. Create searchbar component, and allChat display panel component.
