+++ chat-HTTP-server/chat-HTTP-server/Service/AuthService/TokenService.cs
@@ -22,8 +22,7 @@ public class TokenService : ITokenService
     public string CreateToken(User user, string role)
     {
         var expiration = DateTime.UtcNow.AddMinutes(ExpirationMinutes);
-        Console.WriteLine("------------------------------");
-        Console.WriteLine(expiration);
+       
         var token = CreateJwtToken(
             CreateClaims(user, role),
             CreateSigningCredentials(),
