+++ chat-HTTP-server/chat-HTTP-server/appsettings.Development.json
@@ -1,8 +1,15 @@
 {
   "Logging": {
     "LogLevel": {
-      "Default": "Warning",
-      "Microsoft.AspNetCore": "Warning"
+      "Default": "Information",
+      "Microsoft": "None",
+      "Microsoft.Hosting.Lifetime": "None"
+    },
+    "AuthController": {
+      "IncludeScopes": true,
+      "LogLevel": {
+        "Default": "Information"
+      }
     }
   }
 }
