+++ chat-HTTP-server/chat-HTTP-server/appsettings.json
@@ -1,11 +1,12 @@
 {
   "Logging": {
     "LogLevel": {
-      "Default": "Warning",
-      "Microsoft": "Warning",
-      "Microsoft.Hosting.Lifetime": "Warning"
+      "Default": "None",
+      "Microsoft": "None",
+      "Microsoft.Hosting.Lifetime": "None"
     },
-    "Console": {
+    "AuthController": {
+      "IncludeScopes": true,
       "LogLevel": {
         "Default": "Information"
       }
