+++ chat-frontend/src/context/activeChatContext.jsx
@@ -1,4 +1,12 @@
 import { createContext, useContext } from "react";
 
-const ActiveChatDataContext = createContext(null);
+export const ActivChatDataContext = createContext(null);
+
+export const useActivChatDataContex = () =>{
+    const context = useContext(ActivChatDataContext);
+    if (!context) {
+        throw new Error("useActiveChatDataContex must be used within a ActiveChatDataContexProvider");
+      }
+      return context;
+}
